+++ include/model/model_zoe.h
@@ -1,6 +1,9 @@
 #ifndef __zoe_model_model_zoe_h
 #define __zoe_model_model_zoe_h
 
+#include <metil.h>
+#include <metil_model/metil_model.h>
+
 void zoe_model_zoe_initialize(
   struct metil* _Nonnull,
   struct metil_model* _Nonnull
