+++ include/textures/zoe_texture_hill_lighting.h
@@ -1,0 +1,14 @@
+#ifndef __zoe_textures_zoe_texture_hill_lighting_h
+#define __zoe_textures_zoe_texture_hill_lighting_h
+
+#include <metil_group.h>
+
+#include <Metal/MTLDevice.h>
+#include <Metal/MTLTexture.h>
+
+id<MTLTexture> _Nonnull zoe_texture_hill_lighting_generate(
+  struct metil_group* _Nonnull,
+  id<MTLDevice> _Nonnull
+);
+
+#endif
