+++ metal/zoe_hill.metal
@@ -1,3 +1,5 @@
+#include <object/object_hill.h>
+
 #include <math_c_pi.h>
 #include <math_c_sine.h>
 
@@ -10,6 +12,7 @@
 struct data_vertex {
   float4 position [[position]];
   float2 position_texture;
+  float2 position_lighting;
   float distance;
   float brightness;
 };
@@ -25,19 +28,19 @@ struct data_vertex {
       metil_renderer_vertex_index_parameter_data_frame
     )
   ]],
-  constant struct metil_renderer_data_object* data_object [[
+  constant struct zoe_data_object_hill* data_object [[
     buffer(
       metil_renderer_vertex_index_parameter_data_object
     )
   ]],
-  unsigned int id_vertex [[vertex_id]]
+  unsigned int index_vertex [[vertex_id]]
 ) {
   struct data_vertex data_vertex;
 
   data_vertex.position = (
     data_object->view_model_matrix_projection *
     positions[
-      id_vertex
+      index_vertex
     ]
   );
 
@@ -53,7 +56,9 @@ struct data_vertex {
         data_object->position.z,
         1.0f
       ) +
-      positions[id_vertex]
+      positions[
+        index_vertex
+      ]
     ),
     metal::float4(
       data_frame->position_player.x,
@@ -76,24 +81,50 @@ struct data_vertex {
 
   data_vertex.position_texture.x = (
     positions[
-      id_vertex
+      index_vertex
     ].x +
     sine_frame
   );
 
   data_vertex.position_texture.y = (
     positions[
-      id_vertex
+      index_vertex
     ].z +
     sine_frame
   );
 
-  return data_vertex;
+  data_vertex.position_lighting.x = (
+    0.0f +
+    (
+      positions[
+        index_vertex
+      ].x /
+      data_object->size.x +
+      0.5f
+    )
+  );
+
+  data_vertex.position_lighting.y = (
+    0.0f +
+    (
+      positions[
+        index_vertex
+      ].z /
+      data_object->size.z +
+      0.5f
+    )
+  ); 
+
+  return (
+    data_vertex
+  );
 }
 
 fragment float4 zoe_hill_fragment(
   struct data_vertex data_vertex [[stage_in]],
-  metal::texture2d<half> texture [[ texture(0) ]]
+  metal::texture2d<half> texture [[ texture(0x00) ]],
+  metal::texture2d<half> texture_secondary [[ texture(0x01) ]],
+  metal::texture2d<half> texture_lighting [[ texture(0x02) ]]
 ) {
   constexpr metal::sampler sampler_texture(
     metal::t_address::repeat,
@@ -101,10 +132,36 @@ fragment float4 zoe_hill_fragment(
     metal::s_address::repeat
   );
 
+  constexpr metal::sampler sampler_texture_secondary(
+    metal::mag_filter::nearest
+  );
+
+  constexpr metal::sampler sampler_texture_lighting(
+    metal::mag_filter::linear
+  );
+
   float4 texture_colour = float4(
     texture.sample(
       sampler_texture,
       data_vertex.position_texture / 1000.0f
+    )
+  );
+
+  float4 texture_secondary_colour = (
+    float4(
+      texture.sample(
+        sampler_texture_secondary,
+        data_vertex.position_lighting
+      )
+    )
+  );
+
+  float4 texture_sample_lighting = (
+    float4(
+      texture_lighting.sample(
+        sampler_texture_lighting,
+        data_vertex.position_lighting
+      )
     )
   );
 
@@ -120,9 +177,9 @@ fragment float4 zoe_hill_fragment(
   );
 
   return float4(
-    texture_colour.r * brightness * data_vertex.brightness,
-    texture_colour.g * brightness * data_vertex.brightness,
-    texture_colour.b * brightness * data_vertex.brightness,
+    texture_colour.r * texture_secondary_colour.x * texture_sample_lighting.x * brightness * data_vertex.brightness,
+    texture_colour.g * texture_secondary_colour.y * texture_sample_lighting.y * brightness * data_vertex.brightness,
+    texture_colour.b * texture_secondary_colour.z * texture_sample_lighting.z * brightness * data_vertex.brightness,
     1.0f
   );
 }
